mush_app/requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
docker-pycreds==0.4.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.104.1
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.0
gast==0.5.4
grpc-google-iam-v1==0.12.7
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
h5py==3.10.0
idna==3.6
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
keras==2.15.0
libclang==16.0.6
loguru==0.7.2
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sentry-sdk==1.37.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-addons==0.22.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
tornado==6.3.3
traitlets==5.13.0
twilio==8.10.2
typeguard==2.13.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
virtualenv==20.24.7
wandb==0.16.0
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.3
